+++ InventoryTools/Logic/Columns/IconColumn.cs
@@ -87,7 +87,7 @@ namespace InventoryTools.Logic.Columns
 
 
         public override string Name { get; set; } = "Icon";
-        public override string RenderName => "";
+        public override bool HideHeaderLabel => true;
         public override float Width { get; set; } = 60.0f;
         public override string HelpText { get; set; } = "Shows the icon of the item, pressing it will open the more information window for the item.";
         public override bool HasFilter { get; set; } = false;