+++ InventoryTools/CHANGELOG.md
@@ -8,6 +8,15 @@ Instead the changelog reader and automation surrounding plugin PRs will add the 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.html).
 
+## [14.0.7] - 2026-03-04
+
+### Fixed
+- Fixed hooks for 7.45
+- Fixed issue with Remove column causing right click issues
+
+### Added
+- Added copy item name button column
+
 ## [14.0.6] - 2026-02-08
 
 ### Fixed